refactor: reorganize assistants/ into components/ and logic/ subdirectories
