+++ frontend/svelte-app/src/lib/components/assistants/assistantFormUtils.svelte.test.js
@@ -6,7 +6,7 @@ import {
 	extractModelsFromConnectorData,
 	extractModelsMetadata,
 	highlightPlaceholders
-} from './assistantFormUtils.svelte.js';
+} from './logic/assistantFormUtils.svelte.js';
