+++ frontend/svelte-app/src/lib/components/assistants/components/AssistantDescriptionField.svelte
