+++ frontend/svelte-app/src/lib/components/assistants/components/AssistantFormHeader.svelte
