+++ frontend/svelte-app/src/lib/components/assistants/components/AssistantNameField.svelte
