+++ frontend/svelte-app/src/lib/components/assistants/components/AssistantSharing.svelte
