+++ frontend/svelte-app/src/lib/components/assistants/components/AssistantSharingModal.svelte
