+++ frontend/svelte-app/src/lib/components/assistants/components/KnowledgeBaseSelector.svelte
