+++ frontend/svelte-app/src/lib/components/assistants/components/RagOptionsPanel.svelte
