+++ frontend/svelte-app/src/lib/components/assistants/components/RubricSelector.svelte
