+++ frontend/svelte-app/src/lib/components/assistants/components/SingleFileSelector.svelte
