+++ frontend/svelte-app/src/lib/components/assistants/logic/assistantFormUtils.svelte.js
