+++ frontend/svelte-app/src/lib/components/assistants/logic/importAssistantValidator.js
